zone/requirements.txt
beanie
betterproto
dataclass-wizard
loguru
pygame-ce
pymunk
pytmx
pyyaml
redis
#-e ../shared